pyproject.toml
[project]
name = "state-management-backend"
version = "0.1.0"
description = "FastAPI backend for state management and hiring funnel charts"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "python-multipart==0.0.6",
    "pydantic>=2.7.4",
    "neo4j>=5.28.1",
    "sentence-transformers>=3.2.1",
    "torch>=2.5.1",
    "python-dotenv>=1.0.1",
    "langgraph>=0.2.54",
    "langchain>=0.3.12",
    "langchain-anthropic>=0.3.3",
]

[tool.uv.sources]
